1) minor bug fix:  changed variable sc in ed_init.f90 to integer, as per .... per its just an integer, not much to it
2) changed the order in which leaf interception is calculated in canopy_derivs_two.  The hybrid scheme needs the information prior to the calculation of evaporation.  This change should have no adverse effects on other integrators, leaf_interception has no dependency of leaf state otherwise.

+++ ED/src/init/ed_init.f90
@@ -78,7 +78,7 @@ subroutine set_site_defprops()
    integer                    :: npoly
    integer                    :: nsite
    integer                    :: k
-   real                       :: sc
+   integer                    :: sc
    real                       :: zmin
    real                       :: fa
    real                       :: fb
